Add plumbing for koans runner
Executing each lesson immediately prevents all code from loading before
the project executes. This makes it hard to maintain. Instead we'd like
to build a list of lessons to be run and then step through them in
order. We can use an Agent to maintain this list and then right before
the program exits, run the lessons (inspired by ExUnit).